requirements.txt
kaggle
matplotlib
opencv-python
pandas
pillow
plotly
python-dotenv
pyyaml
scikit-learn
streamlit
torch
torchvision
ultralytics
